Added instructions for installing on NixOS 18.09
Hi there, I've tested to make sure these instructions work in NixOS 18.09, which is still in beta. Unfortunately, the latest versions of Elixir and OTP are not available in the current stable version of NixOS, 18.03. It's also possible to use this shell code with the unstable channel, but that's, well unstable. 

These instructions work for me in multiple window managers and a basic session with nothing but xterm. It's possible these instructions may not work on all machines or installations, but I'd be happy to provide more detail and support for others trying to get Scenic installed on NixOS. 

Thanks for building something incredible.