Add live-reload to new_example
The heavy lifting is done by exsync which supports both Mac and Linux
https://github.com/falood/exsync

Architecture

When exsync detects a file system change it recompiles the changed filed into
the BEAM. After this recompilation sequence finishes, it calls the configured
callback function which is handled in the `Nav` component. The `Nav` component
then restarts scenic by killing it's root process.

If there was an easy/non-invasive way to get the current scene then we could
restart just that instead of the entire process tree.

Note: currently relies on an unreleased version of exsync.