Update graph in `handle_info`
This isn't necessary for the example, but I was using this example as the basis for a project.  When I put in two sensors it was flashing back and forth between the two.  It took me a bit to realize that it was because it was always based on the graph from `init`.  This will hopefully prevent other people from having the same confusion.