scenic.new.nerves: Misc updates to erlinit.config
Since the `erlinit.config` is overridden to redirect output to serial,
this pulls in an improvement to board identification and support for
recovering logs from a previous boot. Neither of these affect Scenic.
Users comparing the erlinit.config file with the official version will
only see the tty change now.